Remove treatment of 'none' from F1 metric
The template is responsible for verbalization and hence it should deal with converting empty label list to 'none' and then from 'none' to an empty list in the post processor.

Until now it was handled in the metric, which complicates it's code, and is error prone, in case a different verbalization is used (e.g. in other languages)

Signed-off-by: Yoav Katz <katz@il.ibm.com>