Apply suggestions from code review
I changed "specified function" to "function specified" (another way of saying the function that is specified) -- because the original problem seemed to stem from using the word "function" at the end of the clause -- the code seemed to flag it as a keyword. I hope it renders better when the word function is in the middle of the text rather than at the end. We'll have to see by looking on the rendered web page after this change is made. Thanks!

Co-authored-by: Yoav Katz <68273864+yoavkatz@users.noreply.github.com>