Update operators.py --- copy edits (grammar, consistency, clarity)
In addition to fixing typos for the word "specialized" and changed "dedicated for" to "dedicated to" -- I also adjusted something that seemed to render differently than planned --

The auto docs said
"additional arguments are field names passed to the function" 

but it showed up on the web page as
"(additional arguments are field names passed to the) - name of function" 

-- it seemed to render this way since the word "function" was at the end.

Incorporating Yoav's new suggested text for clarity, I also switched the final words so that it didn't end with the word function, which seemed to be picked up there by the parser.

I hope this renders better. We can check by looking at the operators.html page on the resulting documentation website after this commit is approved, and if it doesn't work tweak from there.

Please review. Thanks!

Signed-off-by: welisheva22 <welisheva22@gmail.com>