Add generic inference engine to allow dynamic selection by the user
The generic inference engine enables users to set the actual inference
engine dynamically via the inference_engine environment variable.
Additionally, the class includes a default parameter, which specifies the
inference engine to use if the environment variable is not set.

For example, when the inference_engine variable is not set, the engine will
 default to ibm_gen_ai.llama_3_8b_instruct. However, if
inference_engine=ollama.llama2 is set, it will switch to the Ollama model.

This update allows lm-eval users to run the UniTXT dataset with LLMaaJ
using any supported inference engine.