Allow a serializer per field in template.
This allows the template to control the formating of each field (e.g. how to enumerate lists).  This is done by allow the serializer field of the template to be a dictionary of field_names to serializer .

Fixed example for rag response generation.

Also fixed some bugs in serializer code.

Signed-off-by: Yoav Katz <katz@il.ibm.com>